Revert "Feature/drf"
