+++ pyproject.toml
@@ -12,7 +12,6 @@ djangorestframework = "^3.13.1"
 django-filter = "^22.1"
 psycopg2 = "^2.9.4"
 dj-rest-auth = {extras = ["with_social"], version = "^2.2.5"}
-django-username-email = "^2.5.3"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.3"
